refactor(api): implement Unicode data proxy routes and schemas
* Added `V1_RAW_ROUTER` for handling proxy requests to Unicode data files.
* Introduced wildcard routes for accessing specific files and their metadata.
* Created schemas for validating responses from the proxy endpoints.
* Implemented tests for various scenarios including error handling and path validation.
